+++ tracking/trackFindingCDC/hough/perigee/tests/Phi0ImpactCurv.test.cc
@@ -91,7 +91,7 @@ namespace {
       B2DEBUG(100, "Tags of the hits");
       for (const CDCRLTaggedWireHit& rlTaggedWireHit : taggedHits) {
         B2DEBUG(100, "    " <<
-                "rl = " << rlTaggedWireHit.getRLInfo() << " " <<
+                "rl = " << static_cast<int>(rlTaggedWireHit.getRLInfo()) << " " <<
                 "dl = " << rlTaggedWireHit->getRefDriftLength());
       }
